backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Redis==0.4.0
Flask-CORS==4.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
SQLAlchemy==2.0.21
redis==5.0.0
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.2
Werkzeug==2.3.7
aiohttp==3.8.6
openai==1.3.0
bleach==6.1.0
docker==6.1.3
Flask-Limiter==3.5.00.1
bcrypt==4.1.2
gunicorn==21.2.0
eventlet==0.35.2
email-validator==2.1.0.post1
pytest==8.2.2
alembic==1.13.2 